requirements.txt
numpy==1.17.4
torch==1.13.0
sentence_transformers==2.2.2
gensim==3.8.1
tqdm==4.64.1
scipy==0.19.1
xxhash==1.0.1